requirements.txt
PyQt5>=5.15.0
google-auth>=2.15.0,<3.0.0
google-auth-oauthlib>=0.8.0,<2.0.0
google-api-python-client>=2.70.0,<3.0.0
cryptography>=3.4.8
requests>=2.25.1
python-dateutil>=2.8.2
Pillow>=10.0.0
cairosvg>=2.7.0